requirements.txt
discord.py==1.7.3
python-decouple==3.6
pandas==1.1.5
detoxify==0.5.0